Frontend Restructure/Component breakdown
